+++ server/controllers/users.js
@@ -1,7 +1,9 @@
 
 
 import moment from 'moment';
-import { validator, validationErrors } from '../helpers/index';
+import {
+    validator, validationErrors, hashPassword, comparePassword, generateToken,
+} from '../helpers/index';
 import db from '../models/db';
 
 const Meetups = {
@@ -17,28 +19,37 @@ const Meetups = {
         if (error) {
             return validationErrors(res, error);
         }
-        const findAllQuery = 'SELECT * FROM users WHERE username = $1 AND password = $2 LIMIT 1';
+        const findAllQuery = 'SELECT * FROM users WHERE username = $1 LIMIT 1';
         try {
-            const { rows } = await db.query(findAllQuery, [req.body.username, req.body.password]);
-            let response = {};
-            if (rows[0]) {
-                response = {
-                    status: 200,
-                    data: [{
-                        id: rows[0].id,
-                        username: rows[0].username,
-                        firstname: rows[0].firstname,
-                        lastname: rows[0].lastname,
-                        othername: rows[0].othername,
-                    }],
-                };
-            } else {
-                response = {
+            const { rows } = await db.query(findAllQuery, [req.body.username]);
+            if (!rows[0]) {
+                return res.status(404).send({
                     status: 404,
                     message: 'Invalid username or password',
-                };
-                return res.send(response);
+                });
             }
+            if (!comparePassword(rows[0].password, req.body.password)) {
+                return res.status(404).send({
+                    status: 404,
+                    message: 'Invalid username or password',
+                });
+            }
+            const role = ((rows[0].isadmin) ? 'admin' : 'standard');
+            const issueToken = generateToken({
+                user: rows[0].id,
+                username: rows[0].username,
+                firstname: rows[0].firstname,
+                lastname: rows[0].lastname,
+                othername: rows[0].othername,
+                email: rows[0].email,
+                phonenumber: rows[0].phonenumber,
+                role,
+            });
+
+            const response = {
+                status: 200,
+                data: [{ issueToken }],
+            };
             return res.send(response);
         } catch (errorMessage) {
             return res.status(400).send({ status: 400, error: errorMessage });
@@ -76,7 +87,9 @@ const Meetups = {
             });
         }
 
-        const text = 'INSERT INTO users (firstname, lastname, othername, email, phonenumber, username, password, registered, isadmin) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9)';
+        const hashedPassword = hashPassword(req.body.password);
+
+        const text = 'INSERT INTO users (firstname, lastname, othername, email, phonenumber, username, password, registered, isadmin) VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9) returning *';
         const values = [
             req.body.firstname,
             req.body.lastname,
@@ -84,23 +97,29 @@ const Meetups = {
             req.body.email,
             req.body.phonenumber,
             req.body.username,
-            req.body.password,
+            hashedPassword,
             moment().format('YYYY-MM-DD'),
             req.body.isadmin,
         ];
         try {
-            await db.query(text, values);
+            const { rows } = await db.query(text, values);
+
+            const role = ((rows[0].isadmin) ? 'admin' : 'standard');
+
+            const issueToken = generateToken({
+                user: rows[0].id,
+                username: rows[0].username,
+                firstname: rows[0].firstname,
+                lastname: rows[0].lastname,
+                othername: rows[0].othername,
+                email: rows[0].email,
+                phonenumber: rows[0].phonenumber,
+                role,
+            });
+
             const response = {
                 status: 200,
-                data: [{
-                    firstname: req.body.firstname,
-                    lastname: req.body.lastname,
-                    othername: req.body.othername,
-                    email: req.body.email,
-                    phoneNumber: req.body.phoneNumber,
-                    username: req.body.username,
-                    isAdmin: req.body.isadmin,
-                }],
+                data: [{ issueToken }],
             };
             return res.send(response);
         } catch (errorMessage) {
@@ -168,6 +187,9 @@ const Meetups = {
      * @returns {object} User object
      */
     async deleteUser(req, res) {
+        if (req.user.role !== 'admin') {
+            return res.status(400).send({ status: 400, error: 'Unauthorized Access' });
+        }
         const text = 'DELETE FROM users WHERE id = $1 returning *';
         try {
             const { rows } = await db.query(text, [req.params.id]);
@@ -211,7 +233,7 @@ const Meetups = {
                     lastname: rows[0].lastname,
                     othername: rows[0].othername,
                     email: rows[0].email,
-                    phoneNumber: rows[0].phoneNumber,
+                    phonenumber: rows[0].phonenumber,
                     username: rows[0].username,
                     isAdmin: rows[0].isadmin,
                 }],
